+++ src/Pegasus/Client/CIMClient.cpp
@@ -846,9 +846,9 @@ Message* CIMClient::_doRequest(
 
     _requestEncoder->enqueue(request);
 
-    Uint32 startMilliseconds = TimeValue::getCurrentTime().toMilliseconds();
-    Uint32 nowMilliseconds = startMilliseconds;
-    Uint32 stopMilliseconds = nowMilliseconds + _timeOutMilliseconds;
+    Uint64 startMilliseconds = TimeValue::getCurrentTime().toMilliseconds();
+    Uint64 nowMilliseconds = startMilliseconds;
+    Uint64 stopMilliseconds = nowMilliseconds + _timeOutMilliseconds;
 
     while (nowMilliseconds < stopMilliseconds)
     {
@@ -856,7 +856,7 @@ Message* CIMClient::_doRequest(
 	// Wait until the timeout expires or an event occurs:
 	//
 
-	_monitor->run(stopMilliseconds - nowMilliseconds);
+	_monitor->run(Uint32(stopMilliseconds - nowMilliseconds));
 
 	//
 	// Check to see if incoming queue has a message
